% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{aussie_births}
\alias{aussie_births}
\title{Australian Birth Times Data}
\format{
A data frame with 44 observations on the following 3 variables.
\itemize{
\item \code{time:} the number of minutes past midnight at which the
  baby was born.
\item \code{sex:} a character variable, the sex of the baby ("girl"
  or "boy").
\item \code{weight:} the weight of the baby in grams.
}
}
\source{
Steele, S. (December 21, 1997), Babies by the Dozen for Christmas:
24-Hour Baby Boom, The Sunday Mail (Brisbane), page 7.
}
\usage{
aussie_births
}
\description{
The baby arrivals data introduced in Chapter 6 of the STAT0002 notes.
Information concerning the 44 babies (18 girls and 26 boys) born in a
24-hour period at the Mater Mothers' Hospital, Brisbane, Australia, on
December 18, 1997.
}
\keyword{datasets}
