% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat0002-package.R
\docType{data}
\name{berkeley}
\alias{berkeley}
\title{Student Admisions at UCL Berkeley}
\format{
An object of class \code{table} of dimension 2 x 2 x 6.
}
\usage{
berkeley
}
\description{
Aggregate data on applicants to graduate school at Berkeley for the six
largest departments in 1973 classified by admission and sex.
}
\details{
This is a copy of the \code{\link[datasets]{UCBAdmissions}} data
  provided in the \code{\link[datasets:datasets-package]{datasets}} package.
}
\keyword{datasets}
