% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{birds}
\alias{birds}
\title{Between-species dominance in birds data}
\format{
A data frame with 10 observations on the following 8 variables.
\itemize{
\item \code{species:} a two-letter abbreviation for the bird species.
\item \code{species_name:} the common bird species name.
\item \code{sci_name:} the scientific bird species name.
\item \code{mass_g:} mean species mass in grams.
\item \code{dom_rank:} dominance rank, the smaller the value the more
  dominant species tends to be over other species.
\item \code{dom_rank_LCI:} lower 95\% confidence limit for the dominance
  rank.
\item \code{dom_rank_UCI:} upper 95\% confidence limit for the dominance
  rank.
}
}
\source{
Data for species identities, mass and dominance rankings from
  Francis et al (2018). S1 dataset. \doi{10.1371/journal.pone.0202152.s002}
}
\usage{
birds
}
\description{
Francis et al. (2018) conducted an experiment to study the competitive
interactions between 10 species of birds.  These are the data presented in
Table 1 of this paper. \doi{journal.pone.0202152.t001}
}
\references{
Francis M.L., Plummer K.E., Lythgoe B.A., Macallan C., Currie T.E.,
  Blount J.D. (2018) Effects of supplementary feeding on interspecific
  dominance hierarchies in garden birds. PLoS ONE 13(9): e0202152.
  \doi{10.1371/journal.pone.0202152}
}
\keyword{datasets}
