% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat0002.R
\docType{data}
\name{blood_types}
\alias{blood_types}
\title{Blood Types}
\format{
A data frame with 8 observations on the following 3 variables.
\itemize{
\item{\code{ABO:}}{ Blood type in the ABO system (A, B or O).}
\item{\code{rhesus:}}{ Blood type in the Rhesus system (Rh+ or Rh-).}
\item{\code{percentage:}}{ Percentage of people in the ABO-Rhesus blood type}.}
}
\usage{
blood_types
}
\description{
Percentages of people in the UK in the 8 main ABO-Rhesus blood groups.
}
\keyword{datasets}
