% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{box_plot}
\alias{box_plot}
\title{Box plots}
\arguments{
\item{type}{an integer between 1 and 9 selecting one of the nine quantile
algorithms detailed in the \strong{Details} section of the
\code{\link[stats]{quantile}} documentation.}
}
\value{
A list with the same contents as described in the \strong{Value}
  section of the \code{\link{boxplot}} documentation.
}
\description{
Produce box-and-whisker plot(s) of the given (grouped) values, with an
option (using the argument \code{type}) to change the estimator used
to estimate the quartiles.  The only argument described below is
\code{type}.  For details of the other arguments see the
\strong{Arguments} section of \code{\link[graphics]{boxplot}}.
}
\details{
See the \strong{Details} section of the \code{\link{boxplot}}
  documentation.
}
\section{\strong{Usage}}{

  See the \strong{Usage} section of the \code{\link[graphics]{boxplot}}
  documentation.
}

\examples{
birth_times <- ox_births[, "time"]
box_plot(birth_times)
box_plot(birth_times, horizontal = TRUE)
}
