% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{boxplot_stats}
\alias{boxplot_stats}
\title{Box Plot Statistics}
\usage{
boxplot_stats(x, coef = 1.5, do.conf = TRUE, do.out = TRUE, type = 6)
}
\arguments{
\item{x}{a numeric vector for which the boxplot will be constructed
(\code{\link{NA}}s and \code{\link{NaN}}s are allowed and omitted).}

\item{coef}{this determines how far the plot whiskers extend out from the
box. If coef is positive, the whiskers extend to the most extreme data
point which is no more than coef times the length of the box away from
the box. A value of zero causes the whiskers to extend to the data
extremes (and no outliers be returned).}

\item{do.conf, do.out}{logicals; if FALSE, the conf or out component
respectively will be empty in the result.}

\item{type}{Argument \code{type} to \code{\link[stats]{quantile}}.}
}
\description{
A copy of \code{\link[grDevices]{boxplot.stats}} that adds the argument
\code{type} to be passed to \code{\link[stats]{quantile}} when calculating
sample quantiles.
}
