% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat0002-package.R
\docType{data}
\name{clouds}
\alias{clouds}
\title{Cloud-seeding data}
\format{
A data frame with 26 observations on the following 2 variables.
\itemize{
\item{\code{unseeded:}}{ the rainfall amount for unseeded clouds.}
\item{\code{seeded:}}{ the rainfall amount for unseeded clouds.}
}
}
\source{
Simpson, J., Olsen, A. , and Eden. J.C. (1975). A Bayesian analysis
of a multiplicative treatment effect in weather modification.
\emph{Technometrics}, \strong{17}, 161-166.
}
\usage{
clouds
}
\description{
The data used in the STAT0002 notes to introduce the idea of transforming
data to approximate symmetry.  This dataset contains the amounts of
rainfall, in acre-feet, produced by 52 clouds.  Half of the clouds were
chosen at random to be seeded with silver nitrate.
}
\keyword{datasets}
