% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat0002.R
\docType{data}
\name{days}
\alias{days}
\title{Mystery data in Exercises 1}
\format{
A numeric vector.  The unit of measurement is days.
}
\usage{
days
}
\description{
Your task in Exercises 1 is to guess what these data represent using only
summary plots and statistics.
}
\keyword{datasets}
