% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_statistics.R
\name{five_number}
\alias{five_number}
\title{Five number summary}
\usage{
five_number(x, type = 6, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector or matrix.}

\item{type}{Argument \code{type} used in the call to
\code{\link{quantile}} to estimate the 25\%, 50\% and 75\% quantiles.}

\item{na.rm}{A logical scalar.  If true, any \code{\link{NA}} and NaNs
are removed from \code{x} before the sample quantiles are computed.}
}
\value{
A numeric vector (if the input was a vector) or matrix (if the input
  was a matrix).
}
\description{
Calculates the five number summary of a vector of data or of each column
of a matrix of data, using the estimators of the lower quartile,
median and upper quartile in the STAT002 notes.
}
\details{
The five number summary contains the sample minimum and maximum and
  estimates of the lower quartile, median and upper quartile, i.e. the
  25\%, 50\% and 75\% quantiles.  These quantiles are estimated using the
  \code{\link[stats]{quantile}} function.  By default, \code{type = 6} is
  used in the call to \code{quantile} in order to use the estimator defined
  in the STAT002 notes.
}
\examples{
birth_times <- ox_births[, "time"]
five_number(birth_times)

# Note: summary() uses type = 7 in the call to quantile()
five_number(birth_times, type = 7)
summary(birth_times)
}
\seealso{
\code{\link[stats]{quantile}} for calculating sample quantiles.
}
