% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{flu}
\alias{flu}
\title{Influenza data}
\format{
A data frame with 494 observations on the following 2 variables.
\itemize{
\item{\code{date:}}{ the date.}
\item{\code{visits:}}{ the number of people visiting their doctor with
symptoms of flu.}}
}
\usage{
flu
}
\description{
The numbers of people (in thousands of people) in the UK visiting their
doctor with symptoms of influenza (flu) during four-weekly time periods
over the time period 28th January 1967 to 13th November 2004.
}
\keyword{datasets}
