% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ftse}
\alias{ftse}
\title{FTSE 100 Share Index}
\format{
A data frame with 1220 observations on the following 2 variables.
\itemize{
\item{\code{date:}}{ the date.}
\item{\code{price:}}{ the closing price on that date.}}
}
\usage{
ftse
}
\description{
Weekly closing prices of the FTSE 100 share index from 2nd April 1984 to
13th August 2007.
}
\keyword{datasets}
