% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{hubble}
\alias{hubble}
\title{Nebulae data from Hubble (1929)}
\format{
A data frame with 24 observations on the following 2 variables.
\itemize{
\item{\code{distance:}}{ the distance, in megaparsecs, of the nebula
  from Earth.}
\item{\code{velocity:}}{ the velocity, in kilometres per second, of the
  nebula relative to Earth.}
}
}
\source{
Hubble, E. (1929) A relation between distance and radial velocity
among extra-galactic nebulae. \emph{Proceedings of the National Academy of
Science}, \strong{15}, 168-173.
}
\usage{
hubble
}
\description{
The data used in Section 9.1 of the STAT0002 notes to introduce simple
linear regression.  This dataset contains estimates of the distance
from Earth and the velocity relative to Earth of 24 nebulae.
}
\keyword{datasets}
