% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{hubble2}
\alias{hubble2}
\title{Nebulae data from Freedman (2001)}
\format{
A data frame with 24 observations on the following 2
variables.
\itemize{
\item \code{distance:} the distance, in megaparsecs, of the nebula
  from Earth.
\item \code{velocity:} the velocity, in kilometres per second, of the
  nebula relative to Earth.
}
}
\source{
The velocity values are from column 7 of Table 5 of Freedman (2001)
and the distance values are the corresponding values from Table 4.
}
\usage{
hubble2
}
\description{
Data that are an update to the Hubble (1929) \code{\link{hubble}}
data. Contains estimates of the distance from Earth and the velocity
relative to Earth of 24 nebulae.
}
\references{
Freedman et al. (2001) Final results from the Hubble space
telescope key project to measure the Hubble constant.
\emph{The Astrophysical Journal}, \strong{553}, 47-72.
}
\keyword{datasets}
