% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leverage_and_influence_movies.R
\name{lev_inf_movie}
\alias{lev_inf_movie}
\title{Leverage and influence movie}
\usage{
lev_inf_movie(association = c("positive", "negative", "none"), n = 25)
}
\arguments{
\item{association}{A character scalar.  Determines the type of association
between (not-outlying) observations: "positive" for positive linear
association; "negative" negative linear association; "none" for no
association.}

\item{n}{An integer scalar.  The size of the sample of (non-outlying)
observations.}
}
\value{
Nothing is returned, only the animation is produced.
}
\description{
A movie to examine the influence of a single outlying observation on a
least squares regression line.
}
\details{
\code{n} pairs of observations are simulated with the property
  that the mean of response variable \eqn{y} is a linear function of the
  values of the explanatory variable \eqn{x}.  These pairs of observations
  are plotted using filled black circles.  An extra observation is plotted
  using a filled red circle.  Initially this observation is placed in the
  middle of the plot.

  Superimposed on the plot are two least squares regression lines:
  one based on all the data (`with outlier') and one in which the
  `red' observation has been removed (`without outlier'.
  Initially these lines coincide.

  The location of the `red' observation can be changed using the
  +/- buttons so that the effect of the position of this observation
  on the `with outlier' line can be seen.

  We see that if the red observation is outlying, that is, it is far
  from the least squares line fitted to the other observations, then
  its \strong{influence} on the least squares regression line depends on
  its x-coordinate.  If its x-coordinate is much larger or smaller than
  the x-coordinate of the other observations (high \strong{leverage}) then
  the influence is higher than if it has a similar x-coordinate to the
  other observations (low \strong{leverage}).  An observation with high
  leverage does not necessarily have high influence: if its y-coordinate
  falls very close to the regression line fitted to the other observations
  then its influence will be low.
}
\examples{
# Load package rpanel
# [Use install.packages("rpanel") if necessary]
library(rpanel)

\dontrun{
lev_inf_movie()
lev_inf_movie(association = "none")
}
}
\seealso{
\code{\link{movies}}: general information about the movies.
}
