% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{lymphoma}
\alias{lymphoma}
\title{T-cell count data}
\format{
A data frame with 42 observations on the following 2 variables.
\itemize{
\item \code{tcell:} a numeric vector giving the number of T-cells.
\item \code{type:} a factor giving the type of lymphoma: Hodgkin's or
  non-Hodgkin's.
}
}
\usage{
lymphoma
}
\description{
The numbers of a certain type of T-cell (a white blood cell) per cubic mm
in blood samples taken from 42 patients in remission from from lymphoma.
21 of the patients had Hodgkin's lymphoma, the other 21 had non-Hodgkin's
lymphoma.
}
\keyword{datasets}
