% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_vs_median_normal_movie.R
\name{mean_vs_median_normal_movie}
\alias{mean_vs_median_normal_movie}
\title{Sample mean vs sample median: normally distributed data}
\usage{
mean_vs_median_normal_movie(
  n = 10,
  delta_n = 1,
  pos = 1,
  envir = as.environment(pos)
)
}
\arguments{
\item{n}{An integer scalar.  The size of the samples drawn from a
standard normal distribution.}

\item{delta_n}{A numeric scalar.  The amount by which the value of the
sample size is increased/decreased after one click of the +/- button.}

\item{pos}{A numeric integer.  Used in calls to \code{\link{assign}}
to make information available across successive frames of a movie.
By default, uses the current environment.}

\item{envir}{An alternative way (to \code{pos}) of specifying the
environment. See \code{\link{environment}}.}
}
\value{
Nothing is returned, only the animation is produced.
}
\description{
A movie to compare the sampling distributions of the sample mean
and sample median based on a random sample of size \eqn{n} from a
standard normal distribution. This movie can also be opened using
\code{smovie::movies()} and selecting \strong{Mean vs median, normal} from
the \strong{Sampling distributions} menu.
This is based on the \code{\link[smovie]{lev_inf}} function.
(If you have not installed the \code{smovie} package then use
\code{install.packages("smovie")} to install it.)
}
\details{
The movie is based on simulating repeatedly samples of size
  \code{n} from a standard normal N(0,1) distribution.  It contains
  three plots.  The top plot contains a histogram of the most recently
  simulated dataset, with the N(0,1) probability density function (p.d.f.)
  superimposed.

  Each time a sample is simulated the sample mean and sample median are
  calculated.  These values are indicated on the top plot as a filled
  circle: a red circle for the sample mean and a blue circle for the
  sample median; and are added to the values plotted in the middle and
  bottom plots.

  The plot in the middle contains a histogram of
  the sample means of \emph{all} the simulated samples.  The p.d.f.s of
  the sampling distributions of the sample mean (thick red line)
  and the sample medians (thin blue line) are superimposed.

  The plot on the bottom contains a histogram of
  the sample medians of \emph{all} the simulated samples.  The p.d.f.s of
  the sampling distributions of the sample mean (thin red line)
  and the sample medians (thick blue line) are superimposed.

  An extra sample is produced by clicking the + button next to
  "number of samples of size n". [Ignore the - button.]
  The value of the sample size can be changed using the +/-
  buttons in the panel.
}
\examples{
mean_vs_median_normal_movie()
}
\seealso{
\code{\link{stat0002movies}}: general information about the movies.

\code{\link{mean_vs_median_t_movie}}: compares the
  sampling distributions of the sample mean and sample median based
  on a random sample of size \eqn{n} from a Student's t distribution.
}
