% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_vs_median_t_movie.R
\name{mean_vs_median_t_movie}
\alias{mean_vs_median_t_movie}
\title{Sample mean vs sample median: Student's t distributed data}
\usage{
mean_vs_median_t_movie(
  n = 10,
  t_df = 2,
  delta_n = 1,
  pos = 1,
  envir = as.environment(pos)
)
}
\arguments{
\item{n}{An integer scalar.  The size of the samples drawn from a
Student's t distribution.}

\item{t_df}{A positive scalar.  The degrees of freedom of the Student's t
distribution (see \code{\link[stats]{TDist}}).}

\item{delta_n}{A numeric scalar.  The amount by which the value of the
sample size is increased/decreased after one click of the +/- button.}

\item{pos}{A numeric integer.  Used in calls to \code{\link{assign}}
to make information available across successive frames of a movie.
By default, uses the current environment.}

\item{envir}{An alternative way (to \code{pos}) of specifying the
environment. See \code{\link{environment}}.}
}
\value{
Nothing is returned, only the animation is produced.
}
\description{
A movie to compare the sampling distributions of the sample mean
and sample median based on a random sample of size \eqn{n} from a
(central) Student's t distribution.
}
\details{
The movie is based on simulating repeatedly samples of size
  \code{n} from a central Student's t distribution with degrees of
  freedom \code{t_df}.
  It contains three plots.
  The top plot contains a histogram of the most recently
  simulated dataset, with the probability density function (p.d.f.)
  of the Student's t distribution with \code{t_df} degrees of freedom
  superimposed.

  Each time a sample is simulated the sample mean and sample median are
  calculated.  These values are indicated on the top plot as a filled
  circle: a red circle for the sample mean and a blue circle for the
  sample median; and are added to the values plotted in the middle and
  bottom plots.

  The plot in the middle contains a histogram of
  the sample means of \emph{all} the simulated samples.

  The plot on the bottom contains a histogram of
  the sample medians of \emph{all} the simulated samples.

  An extra sample is produced by clicking the + button next to
  "number of samples of size n". [Ignore the - button.]
  The value of the sample size can be changed using the +/-
  buttons in the panel.
}
\examples{
mean_vs_median_t_movie(t_df = 2)
}
\seealso{
\code{\link{stat0002movies}}: general information about the movies.

\code{\link{mean_vs_median_normal_movie}}: compares the
  sampling distributions of the sample mean and sample median based
  on a random sample of size \eqn{n} from a standard normal distribution.
}
