% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_pdf_movie.R
\name{normal_pdf_movie}
\alias{normal_pdf_movie}
\title{Normal p.d.f. movie}
\usage{
normal_pdf_movie(
  starting_mean = 0,
  starting_var = 1,
  delta_mean = 0.1,
  delta_var = 0.1
)
}
\arguments{
\item{starting_mean}{A numeric scalar.  The value of the mean used
to produce the first plot of the p.d.f.}

\item{starting_var}{A positive numeric scalar.  The value of the variance
used to produce the first plot of the p.d.f.}

\item{delta_mean}{A numeric scalar.  The amount by which the value of the
mean is increased/decreased after one click of the +/- button for the
mean.}

\item{delta_var}{A numeric scalar.  The amount by which the value of the
variance is increased/decreased after one click of the +/- button for the
variance.}
}
\value{
Nothing is returned, only the animation is produced.
}
\description{
A movie to show how the probability density function (p.d.f.) of the normal
distribution varies as the mean and/or variance of the distribution
are changed.
}
\details{
The values of the mean and/or variance can be changed using the
  +/- buttons in the panel.
  For the purposes of this movie, the variance cannot be reduced below a
  value of 0.1.
}
\examples{
normal_pdf_movie()
normal_pdf_movie(delta_var = 1)
}
\seealso{
\code{\link{stat0002movies}}: general information about the movies.

\code{\link{normal_areas_movie}}: normal central probability areas
  movie.
}
