% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_sampling_distribution_movie.R
\name{normal_sampling_distns_movie}
\alias{normal_sampling_distns_movie}
\title{Normal sampling distributions movie}
\usage{
normal_sampling_distns_movie(starting_n = 30, delta_n = 1, mu = 0, sigma = 1)
}
\arguments{
\item{starting_n}{A numeric scalar.  The value of the sample size used
to produce the first plot of the p.d.f.}

\item{delta_n}{A numeric scalar.  The amount by which the value of the
sample size is increased/decreased after one click of the +/- button.}

\item{mu}{A numeric scalar.  The mean of the normal distribution from
which the random sample are drawn.}

\item{sigma}{A numeric scalar.  The standard deviation of the normal
distribution from which the random sample are drawn.}
}
\value{
Nothing is returned, only the animation is produced.
}
\description{
A movie to show how the sampling distributions of the sample mean
and sample variance based on a random sample from a normal distribution
depend on the size \eqn{n} of the sample.
}
\details{
The movie is based on two plots.  The top plot shows the
  probability density function (p.d.f.) of the sampling distribution of the
  sample mean, that is, a N(\eqn{\mu}, \eqn{\sigma}^2 / n) distribution.
  The bottom plot contains the p.d.f. of the sampling distribution of the
  sample variance, that is, a gamma distribution
  (\code{\link[stats]{GammaDist}}) with shape parameter
  \eqn{(n - 1) / 2} and rate parameter \eqn{(n - 1) / 2\sigma^2}.

  The number of samples simulated is increased

  The value of the sample size can be changed using the +/-
  buttons in the panel.  For the purposes of this movie the value of
  \eqn{n} cannot exceed 100.
}
\examples{
# Load package rpanel
# [Use install.packages("rpanel") if necessary]
library(rpanel)

\dontrun{
normal_sampling_distns_movie()
}
}
\seealso{
\code{\link{movies}}: general information about the movies.
}
