% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat0002.R
\docType{data}
\name{ox_births}
\alias{ox_births}
\title{Oxford Birth Times}
\format{
A data frame with 95 observations on the following 2 variables.
\itemize{
\item{\code{day:}}{ Day on which the woman arrived.}
\item{\code{time:}}{ Time (hours) spent on delivery suite.}}
}
\source{
Davison, A. C. (2003) Statistical Models. Cambridge University Press.
  Page 18.
}
\usage{
ox_births
}
\description{
Times spent in the delivery suite by 95 women giving birth at the John
Radcliffe Hospital, Oxford. These data have been taken from the SMPracticals
package to which they were provided by Ethel Burns.
}
\keyword{datasets}
