% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ox_births_simulation_movie.R
\name{ox_births_movie}
\alias{ox_births_movie}
\title{Oxford Birth Times Simulation Movie}
\usage{
ox_births_movie(starting_n = 100, delta_n = 1000)
}
\arguments{
\item{starting_n}{A numeric scalar.  The size of the first sample to be
simulated.}

\item{delta_n}{A numeric scalar.  The amount by which the sample size
is increased (or decreased) after one click of the + (or -) button
in the parameter window.}
}
\value{
Nothing is returned, only the animation is produced.
}
\description{
A movie to illustrate that as we increase the size of a sample simulated
from a continuous probability distribution a histogram of these data
gives a closer approximation to the underlying probability density
function (p.d.f.).
}
\details{
\code{ox_births_movie} first fits a
  \href{https://en.wikipedia.org/wiki/Gamma_distribution}{gamma distribution}
  (see \code{\link[stats]{GammaDist}}) to the \code{time} data in the
  \code{\link{ox_births}} data.  Then we simulate samples (using
  \code{\link[stats:GammaDist]{rgamma}}) from the fitted gamma distribution and produce a
  histogram of the simulated data.  The p.d.f. of the fitted gamma distribution
  is superimposed on the histogram.

  We will not study the gamma distribution in STAT0002, but you will encounter
  it in STAT0003 if you take that next term.
}
\examples{
# Load package rpanel
# [Use install.packages("rpanel") if necessary]
library(rpanel)

\dontrun{
ox_births_movie()
}
}
\seealso{
\code{\link{movies}}: general information about the movies.
}
