% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot.poisson_process}
\alias{plot.poisson_process}
\title{Plot diagnostics a \code{poisson_process} object}
\usage{
\method{plot}{poisson_process}(x, y, ...)
}
\arguments{
\item{x}{an object of class "poisson_process", a result of a call to
\code{\link{poisson_process_sim}}.}

\item{y}{Not used.}

\item{...}{Additional arguments passed to
\code{\link[graphics:plot.default]{plot}}}
}
\value{
Nothing is returned.
}
\description{
\code{plot} method for class "poisson_process".
}
\examples{
sim1 <- poisson_process_sim(lambda = 2, hours = 24)
plot(sim1)
sim2 <- poisson_process_sim(lambda = 2, n_events = 50)
plot(sim2, pch = 4, xlab = "event times")
}
\seealso{
\code{\link{poisson_process_sim}}.
}
