% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_process_check_movie.R
\name{poisson_process_check}
\alias{poisson_process_check}
\title{Graphical checking of Poisson process properties}
\usage{
poisson_process_check(
  user_data = NULL,
  total_time = NULL,
  intervals = 1,
  unif_bins = NULL,
  exp_bins = NULL
)
}
\arguments{
\item{user_data}{A numeric vector of arrival times of events.}

\item{total_time}{A positive numeric scalar.  The number of time units
(for example hours) over which the events in \code{user_data}
were recorded.  Must be no smaller than the largest value in
\code{user_data}.}

\item{intervals}{An integer scalar.  Must not be smaller than 1.
The number of intervals into which to split the time interval
(0, \code{total_time}) for the purposes of checking the property
that the number of events that occur in an interval of fixed width
has a Poisson distribution.  If \code{intervals} is not an integer
then it will be rounded to the nearest integer.}

\item{unif_bins}{A positive integer scalar.  The number of bins to use in
the histogram in the "times at which events occur" described in
\strong{Details}.  If this is not supplied then the default
number of bins used in \code{\link[graphics]{hist}} is used.}

\item{exp_bins}{A positive integer scalar.  The number of bins to use in
the histogram in the "times between events" described in
\strong{Details}.  If this is not supplied then the default
number of bins used in \code{\link[graphics]{hist}} is used.}
}
\value{
Nothing is returned, only the animation is produced.
}
\description{
A movie to perform informal graphical checks of whether event arrival times
are consistent with arising from a (one-dimensional, homogeneous)
Poisson process.
}
\details{
This movie contains two displays of plots: one plot on the top and
  one of three other plots on the bottom.  The rate, \eqn{\lambda} say, of
  the supposed Poisson process is estimated by the sample mean number of
  events over the \code{n_intervals}, that is,
  \code{length(user_data)} / \code{intervals}.  Let's call this estimate
  \eqn{m}.

  The type of plot that appears in the bottom of the display depends
  on the radio click by the user. The choices are
  \itemize{
  \item{"none" }{Nothing is plotted}
  \item{"numbers of events in each hour" }{A barplot (the red bars) giving
    the proportions of the hours for which there are 0, 1 , 2, ... events.
    Also included are black bars showing the p.m.f. of a
    Poisson(\eqn{m}) random variable.}
  \item{"times between events" }{A histogram (with red rectangles) of
    the sample times between events, with the p.d.f. of an
    exponential(\eqn{m}) random variable superimposed.}
  \item{"times at which events occur" }{A histogram (with red rectangles)
    of the simulated event times, with the p.d.f. of a
    uniform(0, \code{hours}) random variable superimposed.}
  }
  The type of plot can be changed by clicking on the appropriate radio
  button.
}
\examples{
# Data in which an event occurs exactly on the hour
poisson_process_check(user_data = (1:24), 24, 24)
poisson_process_check(user_data = (1:24), 24, 24, exp_bins = 100)
}
\seealso{
\code{\link{movies}}: general information about the movies.

\code{\link{poisson_process_movie}}: for similar plots based on
  data simulated from a Poisson process.
}
