% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_process_movie.R
\name{poisson_process_movie}
\alias{poisson_process_movie}
\title{Poisson process movie}
\usage{
poisson_process_movie(
  lambda = 1,
  hours = 24,
  pos = 1,
  envir = as.environment(pos)
)
}
\arguments{
\item{lambda}{A positive numeric scalar.  The rate of the Poisson process.
[\code{lambda} must not exceed 200 because the plots in the movie
are not designed to work for larger values of \code{lambda}.]}

\item{hours}{A positive integer scalar.  The number of hours for which to
simulate a Poisson process of rate \code{lambda} events per hour.
For the purposes of this movie \code{hours} must not be smaller than 1.}

\item{pos}{A numeric integer.  Used in calls to \code{\link{assign}}
to make information available across successive frames of a movie.
By default, uses the current environment.}

\item{envir}{An alternative way (to \code{pos}) of specifying the
environment. See \code{\link{environment}}.}
}
\value{
Nothing is returned, only the animation is produced.
}
\description{
A movie to illustrate that if events arrive in a
(one-dimensional, homogeneous) Poisson process of
rate \eqn{\lambda} per hour then (a) the number of events that arrive
during a fixed interval of length \eqn{t} hours has a Poisson
distribution with mean \eqn{\lambda t}; (b) the time between successive
events has an exponential distribution with mean \eqn{1/\lambda}; (c)
the arrival times are an (ordered) random sample from a uniform
distribution on the interval \eqn{(0,t)}.  Data are simulated from
a Poisson process to illustrate this.  Alternatively, the use may supply
their own arrival time data in order to assess, informally using graphs,
whether or not these data seem consistent with arising from a Poisson
process.
}
\details{
This movie contains two displays of plots: one plot on the top and
  one of three other plots on the bottom.

  Data are (repeatedly) simulated from a Poisson process of rate
  \eqn{\lambda} events per hour occur during the time interval
  (0, \code{hours}) hours.  The total numbers of events that occur in each
  hour are also displayed on the plot.  Each time the button "simulate
  another sequence of events" is clicked a new set of simulated events
  is produced.

  The type of plot that appears in the bottom of the display depends
  on the radio button clicked by the user. The choices are
  \itemize{
  \item{"none" }{Nothing is plotted}
  \item{"numbers of events in each hour" }{A barplot (the red bars) giving
    the proportions of the hours for which there are 0, 1 , 2, ... events.
    Also included are black bars showing the p.m.f. of a
    Poisson(\eqn{\lambda}) random variable.}
  \item{"times between events" }{A histogram (with red rectangles) of
    the simulated times between events, with the p.d.f. of an
    exponential(\eqn{\lambda}) random variable superimposed.}
  \item{"times at which events occur" }{A histogram (with red rectangles)
    of the simulated event times, with the p.d.f. of a
    uniform(0, \code{hours}) random variable superimposed.}
  }
  Each time the "simulate another sequence of events" button is clicked
  [or the currently-selected radio button is clicked again] then a new
  set of events is simulated and these event are \strong{added} to
  the current collection of simulated events.

  \strong{Note:} During the early stages of the simulations the
    heights of the black bars in the "numbers of events in each hour"
    plot may be incorrect because they extend beyond the plot region.
}
\examples{
# Load package rpanel
# [Use install.packages("rpanel") if necessary]
library(rpanel)

\dontrun{
poisson_process_movie(lambda = 2)
poisson_process_movie(lambda = 10)
poisson_process_movie(lambda = 0.5)
}
}
\seealso{
\code{\link{movies}}: general information about the movies.

\code{\link{poisson_process_check}}: for similar plots based on
  data supplied by a user.
}
