% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_process_simulation.R
\name{poisson_process_sim}
\alias{poisson_process_sim}
\title{Simulation of a Poisson process}
\usage{
poisson_process_sim(lambda = 1, hours = 24, n_events = NULL)
}
\arguments{
\item{lambda}{A positive numeric scalar.  The rate of the Poisson process.}

\item{hours}{A positive integer scalar.  The number of hours for which to
simulate a Poisson process of rate \eqn{\lambda} events per hour.}

\item{n_events}{A positive integer scalar.  The number of events to
simulate.}
}
\value{
A numeric vector containing the (ordered, smallest to largest)
  times at which the events occur.  The returned object has class
  "poisson_process".
}
\description{
Simulates event times from a (one-dimensional, homogeneous) Poisson process
of rate \eqn{\lambda} per hour.  The user has the options to simulate events
over a fixed time period of \code{hours} hours or to simulate a fixed number
\code{n_events} of events.
}
\details{
If \code{n_events} is supplied then exactly \code{n_events} are
  simulated and \code{hours} has no effect.  If \code{n_events} is not
  supplied then events are simulated over the time interval
  (0, \code{hours}).  If no events occur in (0, \code{hours}) then
  the value \eqn{-1} is returned.
}
\examples{
sim1 <- poisson_process_sim(lambda = 2, hours = 24)
plot(sim1)

sim2 <- poisson_process_sim(lambda = 2, n_events = 50)
plot(sim2)
}
