% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening_test.R
\name{print.screening_test}
\alias{print.screening_test}
\title{Prints a \code{screening_test} object}
\usage{
\method{print}{screening_test}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{an object of class "screening_test", a result of a call to
\code{\link{screening_test}}.}

\item{digits}{An integer.  The number of significant digits to include in
the printed probabilities.  Passed to \code{\link{format}}.}

\item{...}{Additional arguments.  None used at present.}
}
\value{
Nothing is returned.
}
\description{
\code{print} method for class "screening_test".
}
\examples{
screening_test(0.1, 0.9, 0.9)
}
\seealso{
\code{\link{screening_test}}.
}
