% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{qqexp}
\alias{qqexp}
\title{Expnential Quantile-Quantile plots}
\usage{
qqexp(
  y,
  ...,
  main = "Exponential QQ plot",
  line = list(col = "black", lty = 1, lwd = 1),
  envelopes = FALSE
)
}
\arguments{
\item{y}{Sample data}

\item{...}{Optional \code{\link[graphics:par]{graphical parameters}}
passed to \code{\link[graphics:plot.default]{plot}}, such as \code{pch},
\code{lty} and \code{lwd}, to control the appearance of the plot.}

\item{main}{A character scalar. The title (if any) to give the plot.}

\item{line}{Determines whether or not a line of equality is superimposed on
the plot.  If a line is required then must be a list, which can contain
\code{\link[graphics:par]{graphical parameters}}
passed to \code{\link[graphics]{abline}}, such as \code{col}, \code{lty}
and \code{lwd} to control the appearance of the line. If \code{line} is
not a list, for example, \code{line = 0}, then no line is superimposed.}

\item{envelopes}{Determines whether or not simulation envelopes should be
added to the plot.  If \code{envelopes = FALSE} then no envelopes are
added. If \code{envelopes} is a positive integer (a common choice is 19)
then simulation envelopes based on this many simulated datasets are added.
The limits of of the envelopes are indicated using short horizontal lines.}
}
\value{
Nothing, just the plot.
}
\description{
Produces a QQ plot to compare ordered sample data to corresponding
quantiles of an exponential distribution fitted to these data.
}
\details{
The rate parameter \eqn{\lambda} of the exponential distribution
  is estimated using \code{1 / mean(y)}.  The ordered sample data are
  plotted against quantiles of this fitted exponential distribution.
  Specifically, the \eqn{i}th smallest sample observation is plotted
  against the \eqn{100 i / (n + 1)\%} theoretical exponential quantile,
  where \eqn{n} is the sample size. The plot is constrained to be square.
}
\examples{
# Australian Birth Times
# Calculate the waiting times until each birth
waits <- diff(c(0, aussie_births[, "time"]))
qqexp(waits, pch = 16, line = list(lty = 2, col = "blue", lwd = 2))
}
\seealso{
\code{\link[stats]{qqnorm}} for a normal QQ plot.
}
