% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochastic_simulation.R
\name{rbinomial}
\alias{rbinomial}
\title{Simulates from a binomial distribution}
\usage{
rbinomial(size, prob)
}
\arguments{
\item{size}{An integer scalar.  The number of trials.}

\item{prob}{A numeric scalar.  The probability of success on each trial.}
}
\value{
A numeric scalar: the simulated number of successes in \code{size}
  trials.
}
\description{
Simulates \strong{one} value from a binomial distribution with parameters
\code{size} (the number of independent Bernoulli trials) and \code{prob}
(the probability of success on any given trial).
}
\details{
Take a look at this function's code in the
{\href{../doc/stat0002-stochastic-simulation-vignette.html}{Stochastic Simulation}}
vignette and see whether you can understand how it works.
}
\examples{
# Simulate one value from a binomial(6, 0.2) distribution.
rbinomial(size = 6, prob = 0.2)
}
\seealso{
\code{link{rbinom}} for the official R function for simulating
  from a binomial distribution.
}
