% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{scatter}
\alias{scatter}
\title{Scatter plot with five number summary}
\usage{
scatter(x, y, ndec = 1, type = 6, na.rm = FALSE, ...)
}
\arguments{
\item{x}{The argument \code{x} of \code{\link[graphics:plot.default]{plot}}.}

\item{y}{The argument \code{x} of \code{\link[graphics:plot.default]{plot}}.}

\item{ndec}{A numeric vector.  The numbers of decimal places to
which to round the five number summaries.  If \code{ndec} is
a scalar then this value is used for both axes.}

\item{type}{Argument \code{type} used in the call to
\code{\link{five_number}} to estimate the 25\%, 50\% and 75\% quantiles.}

\item{na.rm}{A logical scalar.  If true, any \code{\link{NA}} and NaNs
are removed before the sample quantiles are computed.}

\item{...}{Further arguments to be passed to \code{plot}.}
}
\value{
Nothing, just the plot.
}
\description{
Produces a scatter plot in which the axes are labelled with
the respective five number summaries of the horizontal and vertical
axis data.
}
\examples{
x <- rnorm(100)
y <- rnorm(100)
scatter(x, y)
}
