% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{scatter_hist}
\alias{scatter_hist}
\title{Scatter plot with marginal histograms}
\usage{
scatter_hist(x, y, xbreaks = NULL, ybreaks = NULL, ...)
}
\arguments{
\item{x}{The argument \code{x} of \code{\link[graphics:plot.default]{plot}}.}

\item{y}{The argument \code{x} of \code{\link[graphics:plot.default]{plot}}.}

\item{xbreaks}{A numeric vector.  Optional argument \code{breaks}
to \code{\link[graphics]{hist}} when plotting the histogram
on the horizontal axis.}

\item{ybreaks}{A numeric vector.  Optional argument \code{breaks}
to \code{\link[graphics]{hist}} when plotting the histogram
on the vertical axis.}

\item{...}{Further arguments to be passed to \code{plot}.}
}
\value{
Nothing, just the plot.
}
\description{
Produces a scatter plot in which the axes are supplemented by
histograms of the marginal horizontal and vertical axis data.
}
\examples{
x <- rnorm(100)
y <- rnorm(100)
scatter_hist(x, y)
}
