% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/USelection_movie.R
\name{scatterplot_movie}
\alias{scatterplot_movie}
\title{US 2000 Presidential Election Movie: straightening scatter plots}
\usage{
scatterplot_movie(
  x,
  y,
  delta_power = 0.1,
  pos = 1,
  envir = as.environment(pos),
  ...
)
}
\arguments{
\item{x, y}{Numeric vectors of the same length.  Pairs of values to plot
using a scatter plot.  The values in \code{x} will be plotted on the
horizontal axis, the values in \code{y} on the vertical axis.}

\item{delta_power}{A numeric scalar.  The amount by which the powers of
\code{x} and \code{y} increase/decrease after one click of a button
in the parameter window.}

\item{pos}{A numeric integer.  Used in calls to \code{\link{assign}}
to make information available across successive frames of a movie.
By default, uses the current environment.}

\item{envir}{An alternative way (to \code{pos}) of specifying the
environment. See \code{\link{environment}}.}

\item{...}{Further arguments to be passed to
\code{\link[graphics:plot.default]{plot}}.}
}
\value{
Nothing is returned, only the animation is produced.
}
\description{
A movie to illustrate the effects of the transformation of variable(s)
on the appearance of a scatter plot, using the 2000 U.S. Presidential
Election data from Florida.
}
\details{
\code{scatterplot_movie} produces a scatter plot of the input
  variables \code{x} and \code{y} which can then be animated by
  transforming \code{x} and/or \code{y} using power transformations.
  [In fact a
  \href{https://en.wikipedia.org/wiki/Power_transform#Box.E2.80.93Cox_transformation}{Box-Cox transformation}
  is used.]
  The power of \code{x} and \code{y} is chosen using a parameter windows
  containing buttons labelled + and -.
  Clicking + increases the power by \code{delta_power} and
  clicking - decreases the power by \code{delta_power}.
}
\examples{
# Proportion of votes gained by Buchanan
pbuch <- 100 * USelection$buch / USelection$tvot

# Produce plot
scatterplot_movie(x = USelection$npop, y = pbuch)
# Change the plotting character
scatterplot_movie(x = USelection$npop, y = pbuch, pch = 16)

# Identify Palm Beach using a different plotting character
county_name <- USelection[, "co_names"]
pb <- which(county_name == "PalmBeach")
my_pch <- rep(16, length(county_name))
my_pch[pb] <- 4
scatterplot_movie(x = USelection$npop, y = pbuch, pch = my_pch)
}
\seealso{
\code{\link{movies}}: general information about the movies.

\code{\link{USelection}}: description of the 2000 U.S.
  Presidential Election data from Florida.
}
