% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening_test.R
\name{screening_test}
\alias{screening_test}
\title{Screening test calculations}
\usage{
screening_test(prior, sensitivity, specificity)
}
\arguments{
\item{prior}{A numeric scalar.  The pre-test probability \eqn{P(D)} that the
person has the disease.  In the absence of specific information this could
be the prevalence of the disease in the population, that is, the
proportion of people who have the disease in the general population of
interest.}

\item{sensitivity}{A numeric scalar.  The conditional probability that a
person who has the disease tests positive.  If \eqn{+} is the event that
the randomly chosen person tests positive then the sensitivity is
\eqn{P(+ \mid D)}.}

\item{specificity}{A numeric scalar.  The conditional probability that a
person who does not have the disease tests negative.  If \eqn{-} is the
event that the randomly chosen person test negative then the sensitivity
is \eqn{P(- \mid {\rm not}D)}{P(- | notD)}.}
}
\value{
A list containing the following components
\itemize{
\item \code{pp} The probability \eqn{P(+)} that the person will test positive.
\item \code{ppv} The positive predictive value. The conditional probability
\eqn{P(D \mid +)} if the person tests positive then they has the disease.
\item \code{npv} The negative predictive value. The conditional probability
\eqn{P({\rm not}D \mid -)}{P(notD | -)} if the person tests negative
then they do not have the disease.
\item \code{prior},\code{sensitivity},\code{specificity} The input values of \code{prior},
\code{sensitivity} and \code{specificity}.
}
}
\description{
Consider a screening test for a certain disease or condition, applied to a
person who has a certain prior, pre-test probability of having the disease.
For such a person this function calculates the probabilities that: the
person has the disease given that they test positive; the person does not
have the disease given that they test negative.
}
\details{
The required probabilities are calculated using the \strong{law of total
probability}
\deqn{P(+) = P(+ \mid D) P(D) + P(+ \mid {\rm not}D) P({\rm not}D)}{%
    P(+) = P(+ | D) P(D) + P(+ | notD) P(notD)}
and \strong{Bayes' theorem}
\deqn{P(D \mid +) = \frac{P(+ \mid D) P(D)}{P(+)}}{%
    P(D | +) = P(+ | D) P(D) / P(+)}
\deqn{P({\rm not}D \mid -) = \frac{P(- \mid {\rm not}D) P({\rm not}D)}{P(-)}.}{%
    P(notD | -) = P(- | notD) P(notD) / P(-).}
}
\examples{
screening_test(prior = 0.1, sensitivity = 0.9, specificity = 0.9)
}
