% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{shuttle}
\alias{shuttle}
\title{Challenger Space Shuttle Disaster Dataset}
\format{
A dataframe with 24 rows and 5 columns:
\itemize{
\item \code{flight}: the flight number.
\item \code{date}: the date.
\item \code{damaged}: the number of O-rings with thermal distress.
\item \code{temperature}: temperature, in degrees Fahrenheit.
\item \code{pressure}: pressure, pounds per square inch.
}

The first 23 rows contain data for test flights.  The last row
contains the data for the ill-fated flight on 28th January 1986,
where the number of damaged 0-rings is missing (NA).
}
\source{
Dalal, S. R, Fowlkes, E. B. and Hoadley, B. (1989)
  Risk Analysis of the Space Shuttle: Pre-Challenger Prediction
  of Failure. \emph{Journal of the American Statistical Association},
  \strong{84}(408), 945-957.
  \url{http://dx.doi.org/10.1080/01621459.1989.10478858}
}
\usage{
shuttle
}
\description{
These data are discussed in the first STAT0002 lecture.
In a paper (Dalal \emph{et al.}, 1989) these data are analysed
with a view to estimating (retrospectively) the probability of a
catastrophic failure of the Challenger space shuttle under the
launch conditions on 28th January 1986.
}
\seealso{
\code{\link{shuttle_sim}}, \code{\link{shuttle_sim_plot}}
}
\keyword{datasets}
