% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuttle_movie.R
\name{shuttle_movie}
\alias{shuttle_movie}
\title{Challenger Space Shuttle Disaster Movie}
\usage{
shuttle_movie(n_reps = 1, pos = 1, envir = as.environment(pos))
}
\arguments{
\item{n_reps}{An integer scalar.  Relevant to \code{shuttle_movie} only.
The number of flights to simulate
for each of the 23 (pre-disaster) temperatures in the real dataset.
For example, \code{n_reps = 10} means that we simulate a dataset of
size 230.}

\item{pos}{A numeric integer.  Used in calls to \code{\link{assign}}
to make information available across successive frames of a movie.
By default, uses the current environment.}

\item{envir}{An alternative way (to \code{pos}) of specifying the
environment. See \code{\link{environment}}.}
}
\value{
Nothing is returned, only the animation is produced.
}
\description{
A movie to illustrate uncertainty in the linear logistic regression curves
fitted to the real space shuttle data.
}
\details{
\code{shuttle_movie} produces an
animated version of the plot produced by \code{\link{shuttle_sim_plot}}.
Calling `shuttle_movie` creates a plot containing the real data
and the logistic curve fitted to the real data and opens a
parameter window containing buttons labelled + and -.
Clicking + adds to the plot a new simulated dataset and the logistic curve
fitted to this dataset.
Clicking - removes the mostly recently added dataset.
}
\examples{
# Movie based on datasets of size 23
shuttle_movie()

# Movie based on datasets of size 230
shuttle_movie(n_reps = 10)
}
\seealso{
\code{\link{stat0002movies}}: general information about the movies.

\code{\link{shuttle_sim}} and \code{\link{shuttle_sim_plot}} simulation
  of fake space shuttle data (from a linear logistic regression model)
  and plotting these data, respectively.
}
