% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuttle_sim.R
\name{shuttle_sim}
\alias{shuttle_sim}
\title{Simulate fake space shuttle data}
\usage{
shuttle_sim(n_sim = 1, temperature = NULL)
}
\arguments{
\item{n_sim}{A integer scalar. The number of fake datasets to simulate.}

\item{temperature}{A numeric vector of launch temperatures.

  If \code{temperature} is not supplied then the 23 launch temperatures
  from the real dataset are used.

  If \code{temperature} is supplied then it must be a vector of length one
  (i.e. a scalar). In this event \code{n_sim} gives the number of
  simulated numbers of damaged O-rings at launch temperature
  \code{temperature}.  If \code{temperature} has length greater than one
  then only the first element of \code{temperature} is used and a warning
  is given.}
}
\value{
The output depends on whether or not \code{temperature} is supplied by
the user.

If \code{temperature} \emph{is} supplied then \code{shuttle_sim}
returns a dataframe with 2 + \code{n_sim} columns.
Column 1 contains the launch temperatures, column 2 contains the numbers
of distressed O-rings in the real data and columns 3 to
2 + \code{n_sim} the \code{n_sim} simulated datasets.

If \code{temperature} \emph{is not} supplied then \code{shuttle_sim}
returns a vector of length \code{n_sim}.
}
\description{
Simulates fake O-ring thermal distress data for Challenger Space Shuttle
launches at different launch temperatures.  The simulated data are based
on a linear logistic regression model fitted to the real data from the 23
(pre-disaster) launches.
}
\details{
The data are simulated from a linear logistic regression model
  fitted to the real (pre-disaster) O-ring distress and launch
  temperature data.  For a given launch temperature \eqn{t} this model
  provides an estimate, \eqn{phat(t)} say, of the probability that an
  O-ring suffers thermal distress.  Then the number of the 6 O-rings
  that suffers from thermal distress is simulated from a
  binomial(6, \eqn{phat(t)}) distribution, under an assumption that
  the fates of the O-rings are independent.  This is repeated for each of
  the launch temperatures in \code{temperatures}.
  For further details see the
  \href{../doc/stat0002-shuttle-vignette.html}{Challenger Space Shuttle Disaster}
  vignette.
}
\examples{
# Simulate 10 fake datasets of size 23, using the real temperatures.
res <- shuttle_sim(n_sim = 10)
res
# Simulate the number of distressed O-rings for 1000 launches at 31 deg F.
res <- shuttle_sim(n_sim = 1000, temperature = 31)
res[1:100]
table(res)
}
\seealso{
The \href{../doc/stat0002-shuttle-vignette.html}{Challenger Space Shuttle Disaster}
  vignette.

\code{\link{shuttle_sim_plot}} for assessing uncertainty concerning
  the modelling of the space shuttle data using simulation.
}
