% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuttle_sim.R
\name{shuttle_sim_hists}
\alias{shuttle_sim_hists}
\title{Space shuttle: uncertainty in estimated probability of O-ring damage}
\usage{
shuttle_sim_hists(x, temps, ...)
}
\arguments{
\item{x}{A 2-column matrix returned from a call to
\code{\link{shuttle_sim_plot}}}

\item{temps}{A numeric vector of temperatures, in degrees Fahrenheit.}

\item{...}{Further arguments to be passed to \code{hist}. [Apart from
\code{probability}, \code{xlab}, \code{ylab} and \code{main},
which are set inside \code{shuttle_sim_hists}.]}
}
\value{
Nothing, just the plot.
}
\description{
Illustrates the uncertainty in the estimated probability of an O-ring
suffering thermal distress for a given launch temperature.
}
\details{
For details of the linear logistic model see
  \href{../doc/stat0002-shuttle-vignette.html}{Challenger Space Shuttle Disaster}
  vignette and for simulation from this model see
  \code{\link{shuttle_sim}}.
}
\examples{
x <- shuttle_sim_plot(n_sim = 1000, plot = FALSE)
shuttle_sim_hists(x, temps = c(31, 50, 65, 80), col = 8)
}
\seealso{
\code{\link{shuttle_sim_plot}} and \code{\link{shuttle_sim}}.

The \href{../doc/stat0002-shuttle-vignette.html}{Challenger Space Shuttle Disaster}
  vignette.
}
