% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuttle_sim.R
\name{shuttle_sim_plot}
\alias{shuttle_sim_plot}
\title{Space shuttle: uncertainty in fitted linear logistic regression curve}
\usage{
shuttle_sim_plot(
  n_sim = 50,
  plot_real_data = TRUE,
  n_reps = 1,
  plot = TRUE,
  ...
)
}
\arguments{
\item{n_sim}{An integer scalar.  The number of fake datasets to simulate,
and hence the number of curves from simulated data to appear in the plot.}

\item{plot_real_data}{A logical scalar.  Should we add to the plot the
real data and the linear logistic curve fitted to the real data?
\code{real_data = TRUE} for ``yes'' and \code{real_data = FALSE}
for ``no''.}

\item{n_reps}{An integer scalar.  The number of flights to simulate
for each of the 23 (pre-disaster) temperatures in the real dataset.
For example, \code{n_reps = 10} means that we simulate a dataset of
size 230.}

\item{plot}{A logical scalar indicating whether or not to produce the plot.}

\item{...}{Further arguments to be passed to the \code{lines} function
used to draw the curves for the simulated datasets.}
}
\value{
A numeric matrix with 2 columns and \code{n_sim} rows.
  Each row contains the estimates of the parameters of the linear
  logistic regression model fitted to a simulated dataset.
  The first column, \code{alpha_hat}, contains the estimates of the
  intercept parameter, the second column, \code{beta_hat},
  the estimates of the slope parameter.
}
\description{
Illustrates the uncertainty in the fitted values of the probability
of O-ring distress for different values of temperature, based on the
linear logistic regression model fitted to the real data.
Fake data from this fitted model are simulated repeatedly using
the function \code{\link{shuttle_sim}}.  The linear logistic model
is fitted to each of these simulated datasets.
A plot is produced containing the real proportions of distressed
O-rings, the logistic curve fitted to the real data and each of the
logistic curves fitted to the fake data.
}
\details{
For details of the linear logistic model see
  \href{../doc/stat0002-shuttle-vignette.html}{Challenger Space Shuttle Disaster}
  vignette and for simulation from this model see
  \code{\link{shuttle_sim}}.
}
\examples{
shuttle_sim_plot(n_sim = 50)
}
\seealso{
The \href{../doc/stat0002-shuttle-vignette.html}{Challenger Space Shuttle Disaster}
  vignette.

\code{\link{shuttle_sim}} for simulating fake space shuttle data.

\code{\link{lines}} for arguments that can be supplied in ....
}
