% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_statistics.R
\name{skewness}
\alias{skewness}
\alias{q_skew}
\alias{skew}
\title{Sample skewness functions}
\usage{
q_skew(x, type = 6, na.rm = FALSE)

skew(x, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector or matrix.}

\item{type}{Relevant to \code{q_skew} only.  Argument \code{type} used in
the call to \code{quantile} to estimate the 25\%, 50\% and 75\% quantiles.}

\item{na.rm}{A logical scalar.  If true, any \code{\link{NA}} and NaNs
are removed from \code{x} before the constituent parts of the sample
skewness are computed.}
}
\value{
A numeric scalar (if the input was a vector) or vector (if the input
  was a matrix).
}
\description{
Calculates sample measures of skewness (the sample quartile skewness or
standardized sample skewness) of a vector of data, or of each
column of a matrix of data, based on the estimators described in the
the STAT002 notes.
}
\details{
See Chapter 2 of the STAT002 notes.

\emph{Sample quartile skewness}.
Let \eqn{q_L}, \eqn{m} and \eqn{q_U} be the sample lower quartile,
  mean and upper quartile respectively.  A measure of skewness often called
  the \emph{quartile skewness} is given by
  \deqn{[ (q_U - m) - (m - qL) ] / (q_U - q_L).}

\emph{Standardized sample skewness}.
Denote a vector of data by \eqn{(x_1, ..., x_n)} and let \eqn{xbar} and
\eqn{s} be the sample mean and sample standard deviation respectively.
The standardized sample skewness is given by
\deqn{(1 / n) \sum (x_i - xbar) ^ 3 / s ^ 3,}
where the summation \eqn{\sum} is over \eqn{i = 1, ..., n}.
}
\examples{
birth_times <- ox_births[, "time"]
skew(birth_times)
q_skew(birth_times)
}
\seealso{
\code{\link[stats]{quantile}} for calculating sample quantiles.

\code{\link{mean}} for the sample mean.

\code{\link[stats]{sd}} for the sample standard deviation.
}
