% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat0002-internal.R
\name{stat0002-internal}
\alias{stat0002-internal}
\alias{is.positiveinteger}
\title{Internal stat0002 functions}
\usage{
is.positiveinteger(x, tol = .Machine$double.eps^0.5)
}
\description{
Internal stat0002 functions
}
\details{
These functions are not intended to be called by the user.
}
\keyword{internal}
