% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat0002.R
\docType{data}
\name{USelection}
\alias{USelection}
\title{The 2000 U.S. Presidential Election in Florida}
\format{
A data frame with 67 observations on 22 variables.
Each row relates to a county in Florida.
\itemize{
\item{Columns 1-4, county identifiers and location: }
  {county number \code{co}, county name \code{co_names},
  latitude in degrees north \code{lat},
  longitude in degrees west \code{lon}.}
\item{Columns 5-12, county demographic variables: }
{population in 1997 \code{npop},
percentage of whites in 1996 \code{whit},
percentage of blacks in 1996 \code{blac},
percentage of Hispanics in 1996 \code{hisp},
percentage of the population aged 65 and over in 1996 \code{o65},
percentage of the population that graduated from high school
(1990 census) \code{hsed},
percentage of the population that graduated from college
(1990 census) \code{coll},
mean personal income (1994) \code{inco}.}
\item{Columns 13-22, numbers of votes for candidates: }{
Bush \code{bush}, Gore \code{gore}, Browne \code{brow},
Nader \code{nade}, Harris \code{harr}, Hagelin \code{hage},
Buchanan \code{buch}, McReynolds \code{mcre}, Phillips \code{phil},
Moorehead \code{moor}}.
\item{Column 23, total number of votes cast: \code{tvot}}.}

For full details see Tables 1 to 4 of
\href{http://dx.doi.org/10.1214/ss/1049993203}{Smith (2002)}.
}
\usage{
USelection
}
\description{
Voting results and demographic data for the state of Florida in the
United States presidential election in the year 2000.
}
\references{
Smith, R. L. (2002) A Statistical Assessment of Buchanan's
  Vote in Palm Beach County, \emph{Statistical Science}, \strong{17(4)},
  441-457, \url{http://dx.doi.org/10.1214/ss/1049993203}.
}
\keyword{datasets}
