% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{world_bank}
\alias{world_bank}
\alias{co2_per_capita}
\alias{gdp_per_capita}
\alias{population_size}
\alias{life_expectancy}
\title{World Bank Data}
\format{
A data frame with 217 rows (one for each country) and one column
  for each year in the data, i.e 55 columns.
  The names of the countries are given in the row names of the dataset,
  e.g. by \code{rownames(popn)}.

An object of class \code{data.frame} with 217 rows and 55 columns.

An object of class \code{data.frame} with 217 rows and 55 columns.

An object of class \code{data.frame} with 217 rows and 55 columns.

An object of class \code{data.frame} with 217 rows and 55 columns.
}
\source{
The World Bank's
\href{https://data.worldbank.org/data-catalog/world-development-indicators}{World Development Indicators database}.
}
\usage{
co2_per_capita

gdp_per_capita

population_size

life_expectancy
}
\description{
Selected annual World Development Indicators provided in the World Bank's
World Development Indicators database for the countries of the world,
1960-2014.  The indicators are
\itemize{
  \item \strong{co2_per_capita}: total carbon dioxide emissions per
    capita, in metric tons per person
  \item \strong{gdp_per_capita}: Gross Domestic Product (GDP) per
    capita, in current US dollars per person
  \item \strong{population_size}: total population size
  \item \strong{life_expectancy}: life expectancy at birth, in years
}
}
\keyword{datasets}
