% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diving.R
\name{diving}
\alias{diving}
\alias{dmean}
\alias{dmedian}
\alias{dmode}
\title{Calculation of diving scores}
\usage{
dmean(x, DD = "DD", scores = paste0("J", 1:7), trim = 2/7, replace = FALSE)

dmedian(x, DD = "DD", scores = paste0("J", 1:7), type = 6, replace = FALSE)

dmode(x, DD = "DD", scores = paste0("J", 1:7), replace = FALSE)
}
\arguments{
\item{x}{A numeric matrix or data frame, ideally of the same format as
\code{\link{daley}}.}

\item{DD}{A character (or numeric) scalar giving the column name (or number)
of \code{x} that contains the DD (degree of difficulty) values of each
dive.  The default is set up to work with the data in
\code{\link{daley}}.}

\item{scores}{A character (or numeric) vector giving the column names (or
numbers) of \code{x} that contain the 7 scores for each dive.  The default
is set up to work with the data in \code{\link{daley}}.}

\item{trim}{Only relevant to \code{divemean}. Passed to \code{\link{mean}}.
The default, \code{trim = 2 / 7}, corresponds to the calculation using in
diving competitions.}

\item{replace}{A logical scalar.  If \code{replace = FALSE} then a vector
of the dive points is returned. If \code{replace = TRUE} then the
\code{DivePoints}, \code{DiveRank}, \code{TotalPoints}, \code{OverallRank}
and \code{PointsBehind} columns are updated in light of the new values of
the dive points and this data frame is returned.  If any of these named
columns do not exist then the returned object is the same as the
\code{replace = FALSE} case.}

\item{type}{Only relevant to \code{divemedian}.  Passed to
\code{\link[stats]{quantile}}.  The default, \code{type = 6}, corresponds
to the method that we have studied in STAT0002.}
}
\value{
A numeric vector dive points if \code{replace = FALSE} or \code{x}
  does not have appropriately named columns.  Otherwise, a dataframe of the
  same structure as \code{x} in which the columns \code{DivePoints},
  \code{DiveRank}, \code{TotalPoints}, \code{OverallRank} and
  \code{PointsBehind} have been updated in light of the new dive points
  values.
}
\description{
Functions to implement four ways to calculate the points awarded for a dive
in a diving competition, including the way that is used in the Olympic
Games.
}
\details{
In the default case, e.g. \code{dmean(daley)}, the 2 smallest and
  2 largest scores of the 7 scores are removed; the sum of the 3 remaining
  scores is calculated; and the result is multiplied by the dive's
  degree of difficulty.

  Another way to think of this is that we take the sample mean of the 3
  middle scores, then multiply by 3, then multiply by the degree of
  difficulty.

  In the \code{dmean(daley, trim = 0)} case we take the sample mean of all
  7 scores, then multiply by 3, then multiply by the degree of difficulty.
  That is, we do not trim the 2 smallest and 2 largest values.

  Similarly, in the \code{dmedian(daley)} and \code{dmode(daley)} cases we
  take the sample median or mode, respectively, of all 7 scores, then
  multiply by 3, then multiply by the degree of difficulty. If there is more
  than one sample mode then we use the sample mean of these modes.
}
\examples{
# The calculation used in competitions (trimmed sample sums)
dmean(daley)

# An equivalent calculation without trimming
dmean(daley, trim = 0)

# An equivalent calculation based on a sample median
dmedian(daley)

# An equivalent calculation based on a sample mode
dmode(daley)
}
\seealso{
\code{\link{daley}} for Tokyo 2020 Olympics Diving Men's
  Individual 10m platform final results
}
