% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{qqexp}
\alias{qqexp}
\title{Expnential Quantile-Quantile plots}
\usage{
qqexp(
  y,
  statistic = c("mean", "median"),
  type = 6,
  envelopes = FALSE,
  ...,
  line = list(col = "black", lty = 1, lwd = 1)
)
}
\arguments{
\item{y}{Sample data}

\item{statistic}{A character scalar. Selects the summary statistic used to
estimate \eqn{\lambda}, either the sample mean or sample median.}

\item{type}{An integer scalar. If \code{statistic = "median"} then
this is passed to \code{\link[stats]{quantile}} to select the type of
sample quantile used to estimate the sample median. The default,
\code{type = 6}, selects the estimator defined in the STAT002 notes.}

\item{envelopes}{Determines whether or not simulation envelopes should be
added to the plot.  If \code{envelopes = FALSE} then no envelopes are
added. If \code{envelopes} is a positive integer (a common choice is 19)
then simulation envelopes based on this many simulated datasets are added.
The limits of of the envelopes are indicated using short horizontal lines.}

\item{...}{Optional arguments to be passed to
\code{\link[graphics:plot.default]{plot}} such as \code{xlab},
\code{ylab}, \code{main} and/or
\code{\link[graphics:par]{graphical parameters}} such as \code{pch},
\code{lty} and \code{lwd}, to control the appearance of the main plot.
To change the appearance of the line of equality use the argument
\code{line}.}

\item{line}{Determines whether or not a line of equality is superimposed on
the plot.  If a line is required then must be a list, which can contain
\code{\link[graphics:par]{graphical parameters}}
passed to \code{\link[graphics]{abline}}, such as \code{col}, \code{lty}
and \code{lwd} to control the appearance of the line. If \code{line} is
not a list, for example, \code{line = 0}, then no line is superimposed.}
}
\value{
The estimate of \eqn{\lambda}.
}
\description{
Produces a QQ plot to compare ordered sample data to corresponding
quantiles of an exponential distribution fitted to these data.
}
\details{
The rate parameter \eqn{\lambda} of the exponential distribution
  is estimated using \code{1/mean(y, na.rm = TRUE)} if
  \code{statistic = "mean"} and
  \code{log(2)/quantile(y, probs = 0.5, na.rm = TRUE)} if
  \code{statistic = "median"}.  The ordered sample data are plotted against
  quantiles of this fitted exponential distribution. Specifically, the
  \eqn{i}th smallest sample observation is plotted against the
  \eqn{100 i / (n + 1)\%} theoretical exponential quantile, where \eqn{n} is
  the sample size. The plot is constrained to be square. A line of equality
  is superimposed on the plot.
}
\examples{
## Australian Birth Times Data

# Calculate the waiting times until each birth
waits <- diff(c(0, aussie_births[, "time"]))

# Estimating lambda using the sample mean
qqexp(waits)

# Change the appearance of the points and line
qqexp(waits, pch = 16, line = list(lty = 2, col = "blue", lwd = 2))

# Add simulation envelopes
qqexp(waits, envelopes = 19)

# Estimating lambda using the sample median
qqexp(waits, statistic = "median", envelopes = 19)
}
\seealso{
\code{\link[stats]{qqnorm}} to produce a normal QQ plot.
}
